apps/ocr/ocr-service/requirements.txt
paddleocr>=2.7.0
paddlepaddle>=2.5.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
